backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
google-generativeai==0.8.3
chromadb==0.5.18
python-multipart==0.0.12
python-dotenv==1.0.1
websockets==13.1
pywebpush==2.0.0
cryptography==43.0.3
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
aiofiles==24.1.0
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
pillow==10.4.0
sentence-transformers==3.2.0
